Add S3ResolvedPackagesCacheStorage
